+++ modules/kubernetes/node-local-dns/variables.tf
@@ -1,1 +1,5 @@
-
+variable "dns_ip" {
+    description = "Central DNS IP"
+    type = string
+    default = "10.0.0.10"
+}