+++ modules/kubernetes/eks-core/modules.tf
@@ -122,12 +122,13 @@ module "ingress_nginx" {
 
   source = "../../kubernetes/ingress-nginx"
 
-  cloud_provider            = "aws"
-  http_snippet              = var.ingress_config.http_snippet
-  linkerd_enabled           = var.linkerd_enabled
-  datadog_enabled           = var.datadog_enabled
-  public_private_enabled    = var.ingress_config.public_private_enabled
-  allow_snippet_annotations = var.ingress_config.allow_snippet_annotations
+  cloud_provider                  = "aws"
+  http_snippet                    = var.ingress_config.http_snippet
+  linkerd_enabled                 = var.linkerd_enabled
+  datadog_enabled                 = var.datadog_enabled
+  public_private_enabled          = var.ingress_config.public_private_enabled
+  allow_snippet_annotations       = var.ingress_config.allow_snippet_annotations
+  deny_default_annotation_blocker = var.ingress_config.deny_default_annotation_blocker
 }
 
 module "ingress_healthz" {