+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+
+- [#571](https://github.com/XenitAB/terraform-modules/pull/571) Add storageClass in AKS to enable StandardSSD_ZRS.
+
 ### Changed
 
 - [#565](https://github.com/XenitAB/terraform-modules/pull/565) [Breaking] Update Ingress Nginx major version.